requirements.txt
fastapi
uvicorn
torch
Pillow
opencv-python
numpy
pandas
requests
mangum